% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sums.R
\name{.preciseSumsPtr}
\alias{.preciseSumsPtr}
\title{Get the function pointers to link in C}
\usage{
.preciseSumsPtr()
}
\value{
list of function pointers
}
\description{
This allows the user to use the C functions in their own C/C++
packages without binary linking to PreciseSums.
}
\examples{
.preciseSumsPtr()
}
\author{
Matthew L. Fidler
}
\keyword{internal}
