% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sums.R
\name{neumaierSum}
\alias{neumaierSum}
\title{Using the Neumaier method, take a more accurate sum}
\usage{
neumaierSum(numbers)
}
\arguments{
\item{numbers}{A vector of numbers to sum.}
}
\value{
Sum of numbers, a bit more accurate than kahanSum
}
\description{
Using the Neumaier method, take a more accurate sum
}
\examples{
sum(c(1,1e100,1,-1e100)) ## Should be 2, gives 0
neumaierSum(c(1,1e100,1,-1e100)) ## Gives 2
}
\references{
\url{https://en.wikipedia.org/wiki/Kahan_summation_algorithm}
}
