% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sums.R
\name{psProd}
\alias{psProd}
\title{Using PreciceSums's default method, take a product}
\usage{
psProd(numbers)
}
\arguments{
\item{numbers}{A vector of numbers to sum.}
}
\value{
Product of numbers
}
\description{
Using PreciceSums's default method, take a product
}
