% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sums.R
\name{psSetProd}
\alias{psSetProd}
\title{Choose the type of product to use in PreciceSums.  These are used in the
PreciceSums \code{prod} blocks}
\usage{
psSetProd(type = c("long double", "double", "logify"))
}
\arguments{
\item{type}{Product to use for \code{prod()} in PreciceSums blocks

\code{long double} converts to long double, performs the
multiplication and then converts back.

\code{double} uses the standard double scale for multiplication.}
}
\value{
nothing
}
\description{
Choose the type of product to use in PreciceSums.  These are used in the
PreciceSums \code{prod} blocks
}
\author{
Matthew L. Fidler
}
