% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sums.R
\name{psSum}
\alias{psSum}
\title{Using PreciceSums's default method, take a sum}
\usage{
psSum(numbers)
}
\arguments{
\item{numbers}{A vector of numbers to sum.}
}
\value{
Sum of numbers
}
\description{
Using PreciceSums's default method, take a sum
}
